+++ core-services/prompt-registry/src/test/java/com/sap/ai/sdk/prompt/registry/spring/SpringAiConverterTest.java
@@ -6,17 +6,18 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
 import com.sap.ai.sdk.core.AiCoreService;
-import com.sap.ai.sdk.prompt.registry.PromptClient;
+import com.sap.ai.sdk.prompt.registry.PromptRegistryClient;
+import com.sap.ai.sdk.prompt.registry.client.PromptTemplatesApi;
 import com.sap.ai.sdk.prompt.registry.model.MultiChatContent;
 import com.sap.ai.sdk.prompt.registry.model.MultiChatTemplate;
 import com.sap.ai.sdk.prompt.registry.model.PromptTemplate;
 import com.sap.ai.sdk.prompt.registry.model.PromptTemplateSubstitutionRequest;
 import com.sap.ai.sdk.prompt.registry.model.PromptTemplateSubstitutionResponse;
 import com.sap.cloud.sdk.cloudplatform.connectivity.DefaultHttpDestination;
-import com.sap.cloud.sdk.cloudplatform.connectivity.HttpDestination;
 import java.util.List;
 import java.util.Map;
 import lombok.val;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.springframework.ai.chat.messages.Message;
@@ -28,12 +29,17 @@ public class SpringAiConverterTest {
   private static final WireMockExtension WM =
       WireMockExtension.newInstance().options(wireMockConfig().dynamicPort()).build();
 
-  private final HttpDestination DESTINATION = DefaultHttpDestination.builder(WM.baseUrl()).build();
-  private final AiCoreService SERVICE = new AiCoreService().withBaseDestination(DESTINATION);
+  private static PromptTemplatesApi client;
+
+  @BeforeAll
+  static void setup() {
+    val destination = DefaultHttpDestination.builder(WM.baseUrl()).build();
+    val service = new AiCoreService().withBaseDestination(destination);
+    client = new PromptRegistryClient(service).prompt();
+  }
 
   @Test
   void testPromptRegistryToSpringAi() {
-    var client = new PromptClient(SERVICE);
     val promptResponse =
         client.parsePromptTemplateByNameVersion(
             "categorization",
@@ -56,7 +62,6 @@ public class SpringAiConverterTest {
 
   @Test
   void testInvalidRoleThrowsException() {
-    var client = new PromptClient(SERVICE);
     val errorPrompt =
         client.parsePromptTemplateByNameVersion(
             "categorization",
@@ -75,7 +80,6 @@ public class SpringAiConverterTest {
 
   @Test
   void testMultiChatTemplateTextContentToSpringAi() {
-    var client = new PromptClient(SERVICE);
     val promptResponse =
         client.parsePromptTemplateByNameVersion(
             "categorization",
@@ -94,7 +98,6 @@ public class SpringAiConverterTest {
 
   @Test
   void testMultiChatTemplateImageContentThrowsException() {
-    var client = new PromptClient(SERVICE);
     val promptResponse =
         client.parsePromptTemplateByNameVersion(
             "categorization",